+++ src/app/(blog)/blog/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { redirect } from "next/navigation";
+
+export default function Blog() {
+  redirect("/artikel");
+}
