+++ src/app/(portfolio)/layout.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "@/styles/globals.css";
+import { ReactNode } from "react";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -9,14 +10,10 @@ export const metadata: Metadata = {
   description: "Blog, Porfolio, Anything about work",
 };
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>{children}</body>
-    </html>
-  );
+type Props = {
+  children: ReactNode;
+};
+
+export default function PortfolioLayout({ children }: Props) {
+  return <body className={inter.className}>{children}</body>;
 }
