+++ src/app/layout.tsx
@@ -1,0 +1,9 @@
+import "@/styles/globals.css";
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <html lang="en">{children}</html>;
+}
